Menus: Fix typo in the class attribute for the hidden `title` field in `Walker_Nav_Menu_Checklist`.
Each item that `Walker_Nav_Menu_Checklist` displays is accompanied by several hidden `<input/>` fields that specify default values for each item when added to a menu. These values are passed in JavaScript to the AJAX call triggered when an item is added to a menu.

The hidden field for the title attribute field incorrectly had an underscore instead of a hyphen. Because of this, it was impossible to supply a default value for the Title Attribute field of a nav menu item.

Props yanngarcia, davidbaumwald.
See #47838.

git-svn-id: https://develop.svn.wordpress.org/trunk@46380 602fd350-edb4-49c9-b593-d223f7449a82